fix: compile error; missing cast

+++ targets/postfix_writer.cpp
@@ -217,7 +217,7 @@ void mml::postfix_writer::do_alloc_node(mml::alloc_node * const node, int lvl) {
 
   auto ref = cdk::reference_type::cast(node->type())->referenced();
   node->argument()->accept(this, lvl);
-  _pf.INT(std::max(1, ref->size())); // void has size 0, but we want to alloc 1 byte for it
+  _pf.INT(std::max((size_t) 1, ref->size())); // void has size 0, but we want to alloc 1 byte for it
   _pf.MUL();
   _pf.ALLOC();
   _pf.SP();
